form utils as a class
